+++ services/web/server/tests/unit/isolated/test_exceptions_handlers.py
@@ -20,12 +20,6 @@ from simcore_service_webserver.exceptions_handlers import (
 from simcore_service_webserver.exceptions_handlers_base import (
     _handled_exception_context_manager,
 )
-from simcore_service_webserver.exceptions_handlers_base_2 import (
-    add_exception_handler,
-    add_exception_mapper,
-    handle_registered_exceptions,
-    register_exception_handlers,
-)
 from simcore_service_webserver.exceptions_handlers_http_error_map import (
     _sort_exceptions_by_specificity,
     create_exception_handler_from_http_error_map,
@@ -187,31 +181,3 @@ async def test_create_decorator_from_exception_handler(
         resp = await _rest_handler(
             make_mocked_request("GET", "/foo?raise=ArithmeticError")
         )
-
-
-def test_it():
-
-    app = web.Application()
-
-    async def my_error_handler(request: web.Request, exc: OneError):
-        return web.HTTPNotFound()
-
-    # create register
-    register_exception_handlers(app)
-
-    # register
-    add_exception_handler(app, OneError, my_error_handler)
-
-    # this is a handler create mapping to status_code and reason
-    add_exception_mapper(app, OtherError, web.HTTPNotFound)
-
-    async def foo():
-        raise OneError
-
-    routes = web.RouteTableDef()
-
-    @routes.get("/home")
-    @handle_registered_exceptions()
-    async def home(_request: web.Request):
-        await foo()
-        return web.HTTPOk()
